Rename model tests
